-add homeListAdapter save instance -add SavedStateHandle for moviedetail

+++ app/src/main/java/com/aydemir/movieapp/data/model/ResponseMovieList.kt
@@ -51,5 +51,5 @@ data class Movie(
     @SerialName("vote_average")
     val voteAverage: Double? = null,
     @SerialName("vote_count")
-    val voteCount: Double? = null
+    val voteCount: Int? = null
 ) : Parcelable
